Text-Summarization_GPT2/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
datasets==4.0.0
dill==0.3.8
einops==0.8.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface==0.0.1
huggingface-hub==0.33.4
idna==3.10
jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
lightning-utilities==0.14.3
loguru==0.7.3
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt-einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
peft==0.16.0
pip==25.1.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
sentry-sdk==2.33.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchmetrics==1.7.4
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
